+++ README.md
@@ -72,7 +72,7 @@
 #### 自动同步
 
 ##### 方案A - 强制远程分支覆盖自己的分支
-1. 参考[这里](https://github.com/lxk0301/jd_scripts/blob/master/backUp/gitSync.md)，安装[pull插件](https://github.com/apps/pull)，并确认此项目已在pull插件的作用下（参考文中1-d）。
+1. 参考[这里](https://github.com/BlueskyClouds/My-Actions/blob/master/backUp/gitSync.md)，安装[pull插件](https://github.com/apps/pull)，并确认此项目已在pull插件的作用下（参考文中1-d）。
 2. 确保.github/pull.yml文件正常存在，yml内上游作者填写正确(此项目已填好，无需更改)。
 3. 确保pull.yml里面是`mergeMethod: hardreset`(默认就是hardreset)。
 4. ENJOY!上游更改三小时左右就会自动发起同步。
@@ -82,7 +82,7 @@
 > 上游变动后pull插件会自动发起pr，但如果有冲突需要自行**手动**确认。
 > 如果上游更新涉及workflow里的文件内容改动，需要自行**手动**确认。
 
-1. 参考[这里](https://github.com/lxk0301/jd_scripts/blob/master/backUp/gitSync.md)，安装[pull插件](https://github.com/apps/pull)，并确认此项目已在pull插件的作用下（参考文中1-d）。
+1. 参考[这里](https://github.com/BlueskyClouds/My-Actions/blob/master/backUp/gitSync.md)，安装[pull插件](https://github.com/apps/pull)，并确认此项目已在pull插件的作用下（参考文中1-d）。
 2. 确保.github/pull.yml文件正常存在，yml内上游作者填写正确(此项目已填好，无需更改)。
 3. 将pull.yml里面的`mergeMethod: hardreset`修改为`mergeMethod: merge`保存。
 4. ENJOY!上游更改三小时左右就会自动发起同步。
